chore: rename json for alluvial/update imports
